BUG_FIX: making the library immune to driver-side errors
If not char is recived but some nises are detected it's possible that RTOF will start counting and 1,5T interrupt will be generated... this lead to situation that 3,5 T interrupt will not be genereated while chawr was not recived and library will stay freezed in same point as in previous commit.

Changes are implemented to avoid this and siwtch off library timeout after reciving 3,5T flag not after 1,5T flag. This way Timeout will cover unwanted behaviour and corelation betwenn RTOF and Timer interrupts